+++ main/repositories/src/main/java/org/formulaone/repository/RaceRepository.java
@@ -12,7 +12,7 @@ import java.util.List;
  * @version $Id$
  * @since 1.0
  */
-public interface RaceRepository extends ReadOnlyRepository<Race, Long>,
+public interface RaceRepository extends ReadOnlyRepository<Race, Integer>,
                                         QueryDslPredicateExecutor<Race> {
 
   // --------------------------------------RESULTS--------------------------------------------
