+++ main/services/src/main/java/org/formulaone/service/dto/ConstructorDto.java
@@ -19,7 +19,7 @@ import javax.validation.constraints.Size;
 public class ConstructorDto {
 
   @JsonIgnore
-  private Long id;
+  private Integer id;
 
   @JacksonXmlProperty(isAttribute = true)
   @NotEmpty
@@ -46,11 +46,11 @@ public class ConstructorDto {
     return new Builder();
   }
 
-  public Long getId() {
+  public Integer getId() {
     return id;
   }
 
-  public void setId(Long id) {
+  public void setId(Integer id) {
     this.id = id;
   }
 
@@ -98,14 +98,14 @@ public class ConstructorDto {
    */
   public static class Builder {
 
-    private Long id;
+    private Integer id;
     private String referenceName;
     private String name;
 
     public Builder() {
     }
 
-    public Builder id(Long id) {
+    public Builder id(Integer id) {
       this.id = id;
       return this;
     }
